updated report download and creation
