+++ deepfence_frontend/apps/dashboard/src/components/scan-configure-forms/ComplianceScanConfigureForm.tsx
@@ -320,7 +320,7 @@ export const ComplianceScanConfigureForm = ({
       };
     });
   });
-  console.log(tabs);
+
   const [defaultTab, setDefaultTab] = useState('');
 
   useEffect(() => {
